commed/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.11
Django==3.2.8
django-allauth==0.45.0
djangorestframework==3.12.4
gunicorn==20.0.4
idna==3.3
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
oauthlib==3.1.1
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==5.3.1
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
Pillow==8.4.0
spacy==3.1.3
django-cors-headers
channels==3.0.4
